+++ src/communication/network_cl.c
@@ -2551,7 +2551,7 @@ net_client_check_log_header (LOGWR_CONTEXT * ctx_ptr,
 	    if (logpg_area == NULL)
 	      {
 		er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE,
-			ER_OUT_OF_VIRTUAL_MEMORY, 1, length);
+			ER_OUT_OF_VIRTUAL_MEMORY, 1, (size_t) length);
 		error = ER_OUT_OF_VIRTUAL_MEMORY;
 	      }
 	    css_queue_receive_data_buffer (rc, logpg_area, length);
