+++ src/connection/server_support.c
@@ -2460,7 +2460,7 @@ css_pack_server_name (const char *server_name, int *name_length)
       if (packed_name == NULL)
 	{
 	  er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE, ER_OUT_OF_VIRTUAL_MEMORY,
-		  1, *name_length);
+		  1, (size_t) (*name_length));
 	  return NULL;
 	}
 
@@ -2532,7 +2532,7 @@ css_add_client_version_string (THREAD_ENTRY * thread_p,
 	    {
 	      er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE,
 		      ER_OUT_OF_VIRTUAL_MEMORY, 1,
-		      strlen (version_string) + 1);
+		      (size_t) (strlen (version_string) + 1));
 	    }
 	}
       else
